Fix compiler warnings and (type) safety issues across codebase. Modernized & improved a lot of code.
- Add explicit static_cast for narrowing conversions
- Replace reinterpret_cast on function pointers with memcpy (fixes UB)
- Modernize deprecated APIs (inet_addr > InetPtonA, GetVersionEx > RtlGetVersion - old one lies about Windows version)
- Add range validation for colors, ports, model IDs, array indices
- Fix signed/unsigned comparison warnings
- Add overflow checks before narrowing 64-bit to 32-bit values
- Fix regex escaping in URL validation
- Add missing return statements and WINAPI calling conventions

Misc Fixes (including what definately caused some bugs):

Fixed some signed vs unsigned comparisons
Fixed a regex typo in URL checking
Added missing return statements
Fixed Windows thread function signatures
Make sure math won't overflow before doing it
Check that big numbers fit before stuffing them into smaller types
Clamp colors to 0-255
Check array indexes before using them
Validate port numbers and model IDs
Added static_cast<> everywhere the compiler complains about size mismatches
Check values are in range before shrinking them (e.g make sure a number fits in a byte before storing it as one)

Function Pointer Loading Fix:

Old way of grabbing functions from DLLs was technically UB
Now uses memcpy trick that's actually legal C++
Added size checks to make sure pointers are the same size before copying

Eliminates -Wconversion, -Wsign-conversion, -Wnarrowing warnings WITHOUT hiding (suppression), so in the most proper of ways.

Note: 1.7 (master) will follow suit.
I did the majority of this work on 1.6 first because it allows me to playtest it on public servers with diverse scripts.

+++ Client/mods/deathmatch/ClientCommands.cpp
@@ -658,7 +658,7 @@ void DumpPlayer(CClientPlayer* pPlayer, FILE* pFile)
 
     for (uiIndex = 0; uiIndex < NUM_PLAYER_STATS; uiIndex++)
     {
-        fprintf(pFile, "Stat [%u] = %f\n", uiIndex, pPlayer->GetStat(uiIndex));
+        fprintf(pFile, "Stat [%u] = %f\n", uiIndex, pPlayer->GetStat(static_cast<unsigned short>(uiIndex)));
     }
 
     fprintf(pFile, "Streamed in: %u\n", pPlayer->IsStreamedIn());
