+++ Client/mods/deathmatch/logic/CClientModel.cpp
@@ -231,7 +231,7 @@ void CClientModel::RestoreDFF(CModelInfo* pModelInfo)
         case eClientModelType::TIMED_OBJECT:
         {
             const auto&    objects = &g_pClientGame->GetManager()->GetObjectManager()->GetObjects();
-            unsigned short usParentID = pModelInfo->GetParentID();
+            unsigned short usParentID = static_cast<unsigned short>(pModelInfo->GetParentID());
 
             unloadModelsAndCallEvents(objects->begin(), objects->end(), usParentID, [usParentID](auto& element) { element.SetModel(usParentID); });
 
@@ -251,7 +251,7 @@ void CClientModel::RestoreDFF(CModelInfo* pModelInfo)
         case eClientModelType::VEHICLE:
         {
             CClientVehicleManager* pVehicleManager = g_pClientGame->GetManager()->GetVehicleManager();
-            unsigned short         usParentID = pModelInfo->GetParentID();
+            unsigned short         usParentID = static_cast<unsigned short>(pModelInfo->GetParentID());
 
             unloadModelsAndCallEvents(pVehicleManager->IterBegin(), pVehicleManager->IterEnd(), usParentID,
                                       [usParentID](auto& element) { element.SetModelBlocking(usParentID, 255, 255); });
